Move function from parsers utils to cmseek parsers
