+++ scanners/cmseek/parser/parser.js
@@ -2,8 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-const { addHostnameOrIpToFindingObject } = require("@securecodebox/parser-sdk-nodejs/parser-utils");
-
 async function parse(findings) {
   let results = []
   // Making sure the CMS is Joomla
@@ -62,4 +60,24 @@ async function parse(findings) {
   // concat all parsed results
   return parsed_vulnerabilities.concat(parsed_backupFiles).concat(parsed_debug_mode_enabled)
 }
+
+function addHostnameOrIpToFindingObject(finding, unidentifiedString) {
+  // this function assumes that unidentifiedString is either an ip or an url/hostname 
+  // checking if a string is a valid url is pretty complicated, so it is only checked if the string is an ip.
+
+  // first capture group is a potential protocol, the second capture group is the ip/hostname, the third capture group is a potential port
+  // example: (ssh://)(1.1.1.1)(:20) or (http://)(google.de)(:80) or just 1.1.1.1 or just google.de
+  let regex = /([a-zA-Z]+:\/\/*)?([^\/:]*)(:\d+)?/;
+  let strippedString = regex.exec(unidentifiedString)[2];
+
+  let isIp = require('net').isIP(strippedString);
+  if (isIp) {
+    finding.ip_address = strippedString;
+  }
+  else {
+    finding.hostname = strippedString;
+  }
+  return finding;
+}
+
 module.exports.parse = parse;
